feat: 모달 기능 수정
- 모달 디자인 수정
- 모달이 이벤트의 옆에 뜨도록 `getBoundingClientRect` 로 위치 파악 및 조정
- `closebutton` 이 정상적으로 동작하도록 기능 수정
- `Dialog` 요소 구조 수정
- `DialogTitle`이 `DialogContent` 내부에 있지 않아 발생하는 에러를 `VisuallyHidden` 도입을 통해 해결
