+++ package.json
@@ -18,8 +18,9 @@
     "@fullcalendar/timegrid": "^6.1.15",
     "@radix-ui/react-aspect-ratio": "^1.1.0",
     "@radix-ui/react-dialog": "^1.1.4",
-    "@radix-ui/react-scroll-area": "^1.2.0",
+    "@radix-ui/react-scroll-area": "^1.2.2",
     "@radix-ui/react-slot": "^1.1.0",
+    "@radix-ui/react-visually-hidden": "^1.1.1",
     "@types/styled-components": "^5.1.34",
     "axios": "^1.7.9",
     "class-variance-authority": "^0.7.0",
